Fixed a few warnings.
Rev: src/constants.c:1.23
Rev: src/constants.h:1.14
Rev: src/pike_memory.c:1.74
Rev: src/pike_memory.h:1.16
Rev: src/program.c:1.255
Rev: src/program.h:1.100

+++ src/constants.h
@@ -5,7 +5,7 @@
 \*/
 
 /*
- * $Id: constants.h,v 1.13 2000/08/09 12:49:27 grubba Exp $
+ * $Id: constants.h,v 1.14 2000/08/10 14:46:16 grubba Exp $
  */
 #ifndef ADD_EFUN_H
 #define ADD_EFUN_H
@@ -61,9 +61,9 @@ struct callable *add_efun2(char *name,
 			    optimize_fun optimize,
 			    docode_fun docode);
 struct callable *add_efun(char *name, c_fun fun, char *type, INT16 flags);
-struct callable *quick_add_efun(char *name, int name_length,
+struct callable *quick_add_efun(char *name, ptrdiff_t name_length,
 				c_fun fun,
-				char *type, int type_length,
+				char *type, ptrdiff_t type_length,
 				INT16 flags,
 				optimize_fun optimize,
 				docode_fun docode);
